% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{DataInfoscaling}
\alias{DataInfoscaling}
\title{Data information for reference samples}
\usage{
DataInfoscaling(data, rho, datatype = "abundance", nT = NULL)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of matrices/data.frames (species by sampling units); data can also be input as a single matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (\code{nT}, see below) must be specified.}

\item{rho}{the sampling fraction can be input as a vector for each assemblage, or specify a single numeric value to apply to all assemblages.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence/occurrence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data in a single matrix/data.frame) a vector of positive integers specifying the number of sampling units in each assemblage. If assemblage names are not specified (i.e., \code{names(nT) = NULL}), then assemblages are automatically named as "Assemblage_1", "Assemblage_2",..., etc.}
}
\value{
a data.frame including basic data information.\cr\cr 
For abundance data, basic information includes assemblage name (\code{Assemblage}), sample size in the reference sample (\code{n}), 
total abundance in the overall assemblage (\code{N}), sampling fraction of the reference sample (\code{rho}), 
observed species richness in the reference sample (\code{S.obs}), 
sample coverage estimates of the reference sample (\code{SC(n)}), sample coverage estimate for twice the reference sample size (\code{SC(2n)}),
the first five species abundance counts (\code{f1}--\code{f5}).\cr
 
For incidence data, the basic information includes assemblage name (\code{Assemblage}), number of sampling units in the reference sample (\code{T}), 
total number of sampling units in the overall assemblage (\code{T*}), sampling fraction of the reference sample (\code{rho}), 
total number of incidences (\code{U}), observed species richness (\code{S.obs}), 
sample coverage estimates of the reference sample (\code{SC(T)}), sample coverage estimate for twice the reference sample size
(\code{SC(2T)}), as well as the first five species incidence frequency counts (\code{Q1}--\code{Q5}).\cr
}
\description{
\code{DataInfoscaling} provides basic data information for diversity based on a reference sample.
}
\examples{
# Taxonomic diversity for abundance data
data(Brazil_rainforest_abun_data)
DataInfoscaling(Brazil_rainforest_abun_data, rho = 0.3, datatype = "abundance")


# Taxonomic diversity for incidence data
data(Fish_incidence_data)
DataInfoscaling(Fish_incidence_data, rho = 0.3, datatype = "incidence_raw")


}
