% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{estimatescaling}
\alias{estimatescaling}
\title{Compute diversity estimates with a particular set of sample sizes/coverages}
\usage{
estimatescaling(
  data,
  rho,
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 50,
  conf = 0.95,
  nT = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of matrices/data.frames (species by sampling units); data can also be input as a single matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (\code{nT}, see below) must be specified.}

\item{rho}{the sampling fraction can be input as a vector for each assemblage, or specify a single numeric value to apply to all assemblages.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{c(0, 1, 2)}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence/occurrence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{base}{selection of sample-size-based (\code{base = "size"}) or coverage-based (\code{base = "coverage"}) rarefaction and extrapolation.}

\item{level}{A numerical vector specifying the particular sample sizes or sample coverages (between 0 and 1) for which diversity estimates (q = 0, 1 and 2) will be computed. \cr
If \code{base = "coverage"} (default) and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample coverage among all samples extrapolated to the maximum sample sizes. \cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample size among all samples extrapolated to the maximum sample sizes. \cr
Specifically, the maximum extrapolation limit is set to double the reference sample size when rho is less than 0.2; triple the reference sample size when rho is between 0.2 and 0.4; and the total number of individuals when rho exceeds 0.4.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data in a single matrix/data.frame) a vector of positive integers specifying the number of sampling units in each assemblage. If assemblage names are not specified (i.e., \code{names(nT) = NULL}), then assemblages are automatically named as "Assemblage_1", "Assemblage_2",..., etc.}
}
\value{
a data.frame of diversity table including the following arguments: (when \code{base = "coverage"})
\item{Assemblage}{the name of assemblage.}
\item{Order.q}{the diversity order of q.}
\item{SC}{the target standardized coverage value.}
\item{m, mT}{the corresponding sample size (or number of sampling units) for the standardized coverage value.}
\item{qTD}{the estimated diversity of order q for the target coverage value. The estimate for complete coverage (when \code{base = "coverage"} and \code{level = 1}, or \code{base = "size"} and \code{level = Inf}) represents the estimated asymptotic diversity.}
\item{Method}{Rarefaction, Observed, or Extrapolation, depending on whether the target coverage is less than, equal to, or greater than the coverage of the reference sample.}
\item{s.e.}{standard error of diversity estimate.}
\item{qTD.LCL and qTD.UCL}{the bootstrap lower and upper confidence limits for the diversity of order q at the specified level (with a default value of 0.95).}
Similar output is obtained for \code{base = "size"}. \cr\cr
}
\description{
\code{estimatescaling} computes diversity (Hill numbers with q = 0, 1 and 2) with a particular set of user-specified levels of sample sizes or sample coverages. If no sample sizes or coverages are specified, this function by default computes diversity estimates for the minimum sample coverage or minimum sample size among all samples extrapolated to double reference sizes.
}
\examples{
# Taxonomic diversity for abundance data with two target coverages (93\% and 97\%)
data(Brazil_rainforest_abun_data)
output_est_cov_abun <- estimatescaling(Brazil_rainforest_abun_data, rho = 0.3, q = c(0, 1, 2), 
                                       datatype = "abundance", base = "coverage", 
                                       level = c(0.93, 0.97))
output_est_cov_abun


# Taxonomic diversity for abundance data with two target sizes (1000 and 2000)
data(Brazil_rainforest_abun_data)
output_est_size_abun <- estimatescaling(Brazil_rainforest_abun_data, rho = 0.3, q = c(0, 1, 2), 
                                        datatype = "abundance", base = "size", 
                                        level = c(1000, 2000))
output_est_size_abun 


# Taxonomic diversity for incidence data with two target coverages (97.5\% and 99\%)
data(Fish_incidence_data)
output_est_cov_inci <- estimatescaling(Fish_incidence_data, rho = 0.3, q = c(0, 1, 2), 
                                       datatype = "incidence_raw", base = "coverage", 
                                       level = c(0.975, 0.99))
output_est_cov_inci


# Taxonomic diversity for incidence data with two target sizes (30 and 40)
data(Fish_incidence_data)
output_est_size_inci <- estimatescaling(Fish_incidence_data, rho = 0.3, q = c(0, 1, 2), 
                                        datatype = "incidence_raw", base = "size", 
                                        level = c(30, 40))
output_est_size_inci


}
