% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTscaling}
\alias{ggiNEXTscaling}
\title{ggplot2 extension for an iNEXTscaling object}
\usage{
ggiNEXTscaling(
  output,
  type = 1:3,
  facet.var = "Assemblage",
  color.var = "Order.q"
)
}
\arguments{
\item{output}{an \code{iNEXTscaling} object computed by \code{iNEXTscaling}.}

\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1}); 
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{facet.var}{create a separate plot for each value of a specified variable: 
no separation (\code{facet.var = "None"}); 
a separate plot for each diversity order (\code{facet.var = "Order.q"}); 
a separate plot for each assemblage (\code{facet.var = "Assemblage"}); 
a separate plot for each combination of diversity order and assemblage (\code{facet.var = "Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var = "None"}); 
use different colors for diversity orders (\code{color.var = "Order.q"}); 
use different colors for assemblages/sites (\code{color.var = "Assemblage"}); 
use different colors for combinations of diversity order and assemblage (\code{color.var = "Both"}).}
}
\value{
a \code{ggplot2} object for sample-size-based rarefaction/extrapolation curve (\code{type = 1}), sample completeness curve (\code{type = 2}), and coverage-based rarefaction/extrapolation curve (\code{type = 3}).
}
\description{
\code{ggiNEXTscaling} is a \code{ggplot} extension for an \code{iNEXTscaling} object to plot sample-size- and coverage-based rarefaction/extrapolation sampling curves along with a bridging sample completeness curve.
}
\examples{
\donttest{
# Plot three types of curves of taxonomic diversity with facet.var = "Assemblage"
# for abundance data with order q = 0, 1, 2
data(Brazil_rainforest_abun_data)
output_TD_abun <- iNEXTscaling(Brazil_rainforest_abun_data, rho = 0.3, 
                               q = c(0, 1, 2), datatype = "abundance")
ggiNEXTscaling(output_TD_abun, facet.var = "Assemblage")


# Plot three types of curves of taxonomic diversity for incidence data with order q = 0, 1, 2
data(Fish_incidence_data)
output_TD_inci <- iNEXTscaling(Fish_incidence_data, rho = 0.3, 
                               q = c(0, 1, 2), datatype = "incidence_raw")
ggiNEXTscaling(output_TD_inci)
}


}
