% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggObsAsyscaling}
\alias{ggObsAsyscaling}
\title{ggplot2 extension for plotting q-profile}
\usage{
ggObsAsyscaling(output)
}
\arguments{
\item{output}{the output of the function \code{ObsAsyscaling}.\cr}
}
\value{
a q-profile based on the observed diversity and the asymptotic diversity estimate.\cr\cr
}
\description{
\code{ggObsAsyscaling} is a \code{ggplot2} extension for an \code{ObsAsyscaling} object to plot q-profile (which depicts the observed diversity and asymptotic diversity estimate with respect to order q) for q between 0 and 2 (in increments of 0.2).\cr\cr 
In the plot of profiles, only confidence intervals of the asymptotic diversity will be shown when both the observed and asymptotic diversity estimates are computed.
}
\examples{
# Plot q-profile of taxonomic diversity for abundance data
# with order q between 0 and 2 (in increments of 0.2 by default).
data(Brazil_rainforest_abun_data)
output_ObsAsy_TD_abun <- ObsAsyscaling(Brazil_rainforest_abun_data, rho = 0.3, 
                                       datatype = "abundance")
ggObsAsyscaling(output_ObsAsy_TD_abun)


# Plot q-profile of taxonomic diversity for incidence data
# with order q between 0 and 2 (in increments of 0.2 by default)
data(Fish_incidence_data)
output_ObsAsy_TD_inci <- ObsAsyscaling(Fish_incidence_data, rho = 0.3, 
                                       datatype = "incidence_raw")
ggObsAsyscaling(output_ObsAsy_TD_inci)


}
