% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFun.R
\name{print.iNEXTscaling}
\alias{print.iNEXTscaling}
\title{Printing iNEXTscaling object}
\usage{
\method{print}{iNEXTscaling}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXTscaling} object computed by \code{iNEXTscaling}.}

\item{...}{additional arguments.}
}
\value{
a list of three objects (see \code{iNEXTscaling} for more details) with simplified outputs and notes.
}
\description{
\code{print.iNEXTscaling}: Print method for objects inheriting from class "iNEXTscaling"
}
